requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.16.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.4
orjson==3.10.18
packaging==25.0
passlib==1.7.4
prompt_toolkit==3.0.51
propcache==0.3.2
pyasn1==0.6.1
pydantic==2.11.5
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
twilio==9.6.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
